+++ Source/GateRimSG1/GR_DefOf.cs
@@ -10,6 +10,7 @@ namespace GateRimSG1
         public static HediffDef SG1_GoauldHostSymbiote;
         public static PawnKindDef SG1_GoauldSymbiote;
         public static PawnKindDef SG1_TokraSymbiote;
+        public static PawnKindDef SG1_TokraVoluntaryHost;
         public static FactionDef SG1_Tokra;
         public static JobDef SG1_GoauldAutonomousImplant;
         public static ThingDef SG1_GoauldRitualBasin;